chore: move Duration to gateway-api/src/duration.rs
