+++ gateway-api/src/duration.rs
