--- gateway-api/src/gateway_api/mod.rs
@@ -1,2 +1,0 @@
-pub mod time_duration;
-pub use time_duration::Duration;
